ExistingUrl: Specify a connect timeout
The default curl connect timeout is 300 seconds. Set it to something
much lower to avoid timeouts elsewhere.

This fixes an issue where a user would enter a url for which the DNS
existed, but which was not reachable. The application seemed to freeze,
while the php-pfm processes were happily waiting for the curl timeout.
After a few refreshes, the php-fpm worker process were exhausted and the
application was no longer usable.