ModelAbstract: Allow transformers to return 0 or empty results
Required for gemstracker Model::setChangeFieldsByPrefix() to work when
no user is logged in: the userid to add is UserLoader::UNKNOWN_USER_ID
in that case, which is 0. When the transformer returned 0, the
column/value was not added to the sql query after processing.
This change ensures that the values 0 and an empty string are not
ignored when they are returned by a transformer.

Change getOnLoad() to keep things identical.