PluginLoader: Check the exact value added to the paths list
This fixes an issue where we would check if a path existed in an array
without its trailing slash, after which we would add it *with* a trailing
slash. The result is that the check would always fail, so the path would
always be added. Depending on how many times this was done, the number
of identical paths in the list could become quite large, causing
troubling slowness later on.